Enable logging other players the first time the addon is loaded.
As of this commit, the console variable "cmbtlog.disableOtherPlayers"
is broken and still logs damage for other players when set to true,
but it stops logging other things like interrupts, so this might cause
some confusion. Turning it off by default on first load.

WildstarLogs turns it on by default on every reloadui anyway.