Cleanup: Simplify broken window resizing code.
This is much better. We rely a lot more on native UI stuff (window
moving, resizing), only listen to WindowMove, and the only 'hack'
is the addition of two invisible 8-pixels wide blockers in the
header top left and top right to prevent people from resizing from
the top of the window.

Maybe sometime soon, I'll find a better way do do the resizing
with the "snap to row height"-trick, that wouldn't break things
on resizing from the top instead of the bottom.

This fixes half of #6, the other half coming in the next commit.